Add basic training (GCN+pool) with batches #4
Added a new network implementation in Net() that basically takes a
typical GCN with two layers and then a fully convolutional layer to emit
the scores after softmax. The graphs after the GCN layers are pooled
with graclus to reduce them to two nodes (one for each class).

Also fixed some mistakes in the visualization routines and batch shapes
(specifically with 'pos' and 'x' for the graphs which should be
[num_nodes, num_dimensions] and [num_nodes, num_node_features]
respectively. That needed some fixes in the batch looping process too to
properly index the positions for visualization.