add argument parsing and timing for training
The Python package argparse has been used to add argument parsing
which are then passed to the train function. For the moment, only
three arguments can be provided: batch size, learning rate, and
the number of training epochs.

Timing for the training loop has been added using the timeit module
which is supposed to return the wall clock time using the default
timer depending on the OS.