add two GCN networks with one and three layers #10
We have added a couple more models to the networks module that
represent another two GCN-based networks this time with one and
three layers instead of two as we did with the baseline. Their
filters are 32 and 32/64/128 features respectively. The goal
was to experiment with an increasing number of layers.