add testing no pooling network which achieves 84% acc. #10
We have added a temporary network under test that does not use
pooling to generate the output predictions. Instead it uses fully
connected layers with a flattened input coming from the previous
convolutional layers. Then the output of such connected layers is
reshaped to the needed output shape. This configuration achieves
the milestone of 84% accuracy.