add testing no pooling network which achieves 88% acc. #10
We have added a temporary testing network which again does not make
use of pooling layers to generate the output prediction. This time,
we have added another GCN layer and increased the final number of
generated features to 32. We have also modified the input and output
features of the fully connected layers consequently. This new impl.
achieves 88% accuracy.