separate training and evaluation scripts and add test split loading
We have separated the training and evaluation scripts into two separate
scripts train.py and evaluate.py. Now the training script saves the
model checkpoint on both train_traintest and train_kfold when the best
validation accuracy is found.

Besides, we have refactored the evaluation routines into a new file in
the utils module.

Furthermore, we have added the option in the BioTacSp dataset to load
and generate a .pt file for either the train or the test set.