add script for train-test with custom CSVs for each split
The new script train_csv.py allows you to choose what train CSVs and
which test CSVs you want to use for training and testing respectively.

The dataset module has been changed as well to incorporate this
possibility so that you can either provide a default split or a custom
CSV list.