Fix bug with test set and add data augmentation with sensors' noise.
Whenever there was no test set specified, some parts of the training loop still tried to access some variables related to it, breaking the code in consequence. Now, all of that is fixed by checking first if a test set was provided in the 'args' variable.

The noise of the sensors are now being used for generating an augmented batch sample. The code is still under development but it is working stable.